TaskActivity becomes CentralHub. RobotDriver is created to take care of driving the robot. Code for moving the robot is moved from CentralHub to RobotDriver. IOIOClass, NavigationService, and IncomingSms are changed accordingly. activity_hub, and activity_driver are the two views of CentralHub and RobotDriver respectively. accessory_filter is created to help AndroidManifest allow the app permission to Open Accessory. strings.xml is updates with new strings. build.gradle is updated with kotlin.
