+++ app/src/main/res/values/strings.xml
@@ -5,6 +5,7 @@
     <string name="direction">Direction</string>
     <string name="motion">Motion</string>
     <string name="distance">Distance</string>
-    <string name="send_sms">Send SMS</string>
+    <string name="start_robotDriver">Drive</string>
+    <string name="go">Go</string>
     <string name="test_twitter">Tweet</string>
 </resources>
