+++ app/src/main/res/xml/accessory_filter.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <usb-accessory model="IOIO" />
+</resources>
